some changes on MyBookingPage : 1. Add Modal 2. button xendit migrate to Modal.jsx
