+++ src/views/MyBookingPage.jsx
@@ -1,164 +1,181 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import HeaderComponent1 from "../components/HeaderComponent1";
-import { useMutation } from "@apollo/client";
-import { xenditPayment } from "../config/query";
-import * as faceapi from "@vladmandic/face-api";
+// import { useMutation } from "@apollo/client";
+// import { xenditPayment } from "../config/query";
+// import * as faceapi from "@vladmandic/face-api";
+
+import Modal from "../components/Modal";
 
 const MyBookingPage = () => {
-  const [modelsLoaded, setModelsLoaded] = React.useState(false);
-  const [captureVideo, setCaptureVideo] = React.useState(false);
-  const [labeledDescriptor, setLabeledDescriptor] = React.useState([]);
-  const [authStatus, setAuthStatus] = React.useState(false);
-  const [xenditPay, { error, loading, data }] = useMutation(xenditPayment);
-  const openNewTab = () => {
-    window.open(data && data?.createInvoice.data?.invoice_url);
-  };
-  useEffect(() => {
-    if (data && !loading) {
-      openNewTab();
-    }
-  }, [data, loading]);
+  // const [modelsLoaded, setModelsLoaded] = React.useState(false);
+  // const [captureVideo, setCaptureVideo] = React.useState(false);
+  // const [labeledDescriptor, setLabeledDescriptor] = React.useState([]);
+  // const [authStatus, setAuthStatus] = React.useState(false);
+  // const [xenditPay, { error, loading, data }] = useMutation(xenditPayment);
+  // const openNewTab = () => {
+  //   window.open(data && data?.createInvoice.data?.invoice_url);
+  // };
 
-  const videoRef = React.useRef();
-  const videoHeight = 480;
-  const videoWidth = 640;
-  const canvasRef = React.useRef();
+  const [modalOn, setModalOn] = useState(false);
 
-  React.useEffect(() => {
-    const loadModels = async () => {
-      const MODEL_URL = process.env.PUBLIC_URL + "/models";
-      Promise.all([
-        faceapi.nets.tinyFaceDetector.loadFromUri(MODEL_URL),
-        faceapi.nets.faceLandmark68Net.loadFromUri(MODEL_URL),
-        faceapi.nets.faceRecognitionNet.loadFromUri(MODEL_URL),
-        faceapi.nets.faceExpressionNet.loadFromUri(MODEL_URL),
-        faceapi.nets.ssdMobilenetv1.loadFromUri(MODEL_URL),
-      ]).then(async (_) => {
-        setModelsLoaded(true);
-        // pake localstorage image url
-        const img = await faceapi.fetchImage(localStorage.getItem("imageUrl"));
-        const detections = await faceapi
-          .detectSingleFace(img)
-          .withFaceLandmarks()
-          .withFaceDescriptor();
-        const newLabeledDescriptor = new faceapi.LabeledFaceDescriptors(
-          localStorage.getItem("username"),
-          [detections.descriptor]
-        );
-        setLabeledDescriptor([...labeledDescriptor, newLabeledDescriptor]);
-      });
-    };
-    loadModels();
-  }, []);
+  // useEffect(() => {
+  //   if (data && !loading) {
+  //     openNewTab();
+  //   }
+  // }, [data, loading]);
 
-  const startVideo = () => {
-    setCaptureVideo(true);
-    navigator.mediaDevices
-      .getUserMedia({ video: { width: 300 } })
-      .then((stream) => {
-        let video = videoRef.current;
-        video.srcObject = stream;
-        video.play();
-      })
-      .catch((err) => {
-        console.error("error:", err);
-      });
-  };
+  // const videoRef = React.useRef();
+  // const videoHeight = 480;
+  // const videoWidth = 640;
+  // const canvasRef = React.useRef();
 
-  const handleVideoOnPlay = () => {
-    let detectedUserName = [];
-    setInterval(async () => {
-      if (canvasRef && canvasRef.current) {
-        canvasRef.current.innerHTML = faceapi.createCanvasFromMedia(videoRef.current);
-        const displaySize = {
-          width: videoWidth,
-          height: videoHeight,
-        };
-        new faceapi.matchDimensions(canvasRef.current, displaySize);
+  // React.useEffect(() => {
+  //   const loadModels = async () => {
+  //     const MODEL_URL = process.env.PUBLIC_URL + "/models";
+  //     Promise.all([
+  //       faceapi.nets.tinyFaceDetector.loadFromUri(MODEL_URL),
+  //       faceapi.nets.faceLandmark68Net.loadFromUri(MODEL_URL),
+  //       faceapi.nets.faceRecognitionNet.loadFromUri(MODEL_URL),
+  //       faceapi.nets.faceExpressionNet.loadFromUri(MODEL_URL),
+  //       faceapi.nets.ssdMobilenetv1.loadFromUri(MODEL_URL),
+  //     ]).then(async (_) => {
+  //       setModelsLoaded(true);
+  //       // pake localstorage image url
+  //       const img = await faceapi.fetchImage(localStorage.getItem("imageUrl"));
+  //       const detections = await faceapi
+  //         .detectSingleFace(img)
+  //         .withFaceLandmarks()
+  //         .withFaceDescriptor();
+  //       const newLabeledDescriptor = new faceapi.LabeledFaceDescriptors(
+  //         localStorage.getItem("username"),
+  //         [detections.descriptor]
+  //       );
+  //       setLabeledDescriptor([...labeledDescriptor, newLabeledDescriptor]);
+  //     });
+  //   };
+  //   loadModels();
+  // }, []);
 
-        const faceMatcher = new faceapi.FaceMatcher(labeledDescriptor, 0.5);
+  // const startVideo = () => {
+  //   setCaptureVideo(true);
+  //   navigator.mediaDevices
+  //     .getUserMedia({ video: { width: 300 } })
+  //     .then((stream) => {
+  //       let video = videoRef.current;
+  //       video.srcObject = stream;
+  //       video.play();
+  //     })
+  //     .catch((err) => {
+  //       console.error("error:", err);
+  //     });
+  // };
 
-        //.findBestMatch(detections[0].descriptor)
+  // const handleVideoOnPlay = () => {
+  //   let detectedUserName = [];
+  //   setInterval(async () => {
+  //     if (canvasRef && canvasRef.current) {
+  //       canvasRef.current.innerHTML = faceapi.createCanvasFromMedia(
+  //         videoRef.current
+  //       );
+  //       const displaySize = {
+  //         width: videoWidth,
+  //         height: videoHeight,
+  //       };
+  //       new faceapi.matchDimensions(canvasRef.current, displaySize);
 
-        const detections = await faceapi
-          .detectAllFaces(videoRef.current)
-          .withFaceLandmarks()
-          .withFaceDescriptors()
-          .withFaceExpressions();
-        //new faceapi.TinyFaceDetectorOptions())
-        // console.log(detections);
-        const resizedDetections = faceapi.resizeResults(detections, displaySize);
-        canvasRef &&
-          canvasRef.current &&
-          canvasRef.current.getContext("2d").clearRect(0, 0, videoWidth, videoHeight);
-        // console.log(resizedDetections);
-        // console.log(resizedDetections);
-        const results = resizedDetections.map((d) => {
-          return faceMatcher.findBestMatch(d.descriptor);
-        });
-        detectedUserName.push(results[0]?.label);
-        // console.log("masuk");
-        // const results = new faceapi.FaceMatcher(labeledDescriptor, 0.6).findBestMatch(detections[0].descriptor);
-        // console.log("keluar");
-        results.forEach((result, i) => {
-          const box = resizedDetections[i].detection.box;
-          const drawBox = new faceapi.draw.DrawBox(box, {
-            label: result.toString(),
-          });
-          drawBox.draw(canvasRef.current);
-          canvasRef &&
-            canvasRef.current &&
-            faceapi.draw.drawFaceExpressions(canvasRef.current, resizedDetections);
-        });
-        console.log(detectedUserName);
-      }
-      // nanti pake localstorage username
-      if (detectedUserName.includes(localStorage.getItem("username"))) {
-        closeWebcam();
-        setAuthStatus(true);
-      }
-    }, 1000);
-  };
+  //       const faceMatcher = new faceapi.FaceMatcher(labeledDescriptor, 0.5);
 
-  const closeWebcam = () => {
-    videoRef.current?.pause();
-    videoRef.current?.srcObject.getTracks()[0].stop();
-    setCaptureVideo(false);
-  };
+  //       //.findBestMatch(detections[0].descriptor)
+
+  //       const detections = await faceapi
+  //         .detectAllFaces(videoRef.current)
+  //         .withFaceLandmarks()
+  //         .withFaceDescriptors()
+  //         .withFaceExpressions();
+  //       //new faceapi.TinyFaceDetectorOptions())
+  //       // console.log(detections);
+  //       const resizedDetections = faceapi.resizeResults(
+  //         detections,
+  //         displaySize
+  //       );
+  //       canvasRef &&
+  //         canvasRef.current &&
+  //         canvasRef.current
+  //           .getContext("2d")
+  //           .clearRect(0, 0, videoWidth, videoHeight);
+  //       // console.log(resizedDetections);
+  //       // console.log(resizedDetections);
+  //       const results = resizedDetections.map((d) => {
+  //         return faceMatcher.findBestMatch(d.descriptor);
+  //       });
+  //       detectedUserName.push(results[0]?.label);
+  //       // console.log("masuk");
+  //       // const results = new faceapi.FaceMatcher(labeledDescriptor, 0.6).findBestMatch(detections[0].descriptor);
+  //       // console.log("keluar");
+  //       results.forEach((result, i) => {
+  //         const box = resizedDetections[i].detection.box;
+  //         const drawBox = new faceapi.draw.DrawBox(box, {
+  //           label: result.toString(),
+  //         });
+  //         drawBox.draw(canvasRef.current);
+  //         canvasRef &&
+  //           canvasRef.current &&
+  //           faceapi.draw.drawFaceExpressions(
+  //             canvasRef.current,
+  //             resizedDetections
+  //           );
+  //       });
+  //       console.log(detectedUserName);
+  //     }
+  //     // nanti pake localstorage username
+  //     if (detectedUserName.includes(localStorage.getItem("username"))) {
+  //       closeWebcam();
+  //       setAuthStatus(true);
+  //     }
+  //   }, 1000);
+  // };
+
+  // const closeWebcam = () => {
+  //   videoRef.current?.pause();
+  //   videoRef.current?.srcObject.getTracks()[0].stop();
+  //   setCaptureVideo(false);
+  // };
+
   // {
   //   "transactionId": null,
   //   "hotelApiId": null,
   //   "accessToken": null,
   //   "price": null
   // }
-  const doXenditPayment = async () => {
-    try {
-      await xenditPay({
-        variables: {
-          hotelApiId: 10391433, // Nanti pake hasil fetch .price NANTI APUS KALO UDAH ADA THANKYOU NEXT
-          transactionId: 8, // dari fetch .transactionId
-          accessToken: localStorage.getItem("token"),
-          price: 320000000, // Nanti pake hasil fetch .price
-        },
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
 
-  const btnPay = () => {
-    if (authStatus) {
-      return (
-        <button
-          className="bg-[#266c6b] hover:bg-[#0d423f] focus:ring-4 focus:ring-blue-300 text-white px-6 py-2 shadow-md rounded-md"
-          style={{ marginLeft: 10 }}
-          onClick={() => doXenditPayment()}
-        >
-          Pay with xendit
-        </button>
-      );
-    }
-  };
+  // const doXenditPayment = async () => {
+  //   try {
+  //     await xenditPay({
+  //       variables: {
+  //         hotelApiId: 10391433, // Nanti pake hasil fetch .price NANTI APUS KALO UDAH ADA THANKYOU NEXT
+  //         transactionId: 8, // dari fetch .transactionId
+  //         accessToken: localStorage.getItem("token"),
+  //         price: 320000000, // Nanti pake hasil fetch .price
+  //       },
+  //     });
+  //   } catch (error) {
+  //     console.log(error);
+  //   }
+  // };
+
+  // const btnPay = () => {
+  //   if (authStatus) {
+  //     return (
+  //       <button
+  //         className="bg-[#266c6b] hover:bg-[#0d423f] focus:ring-4 focus:ring-blue-300 text-white px-6 py-2 shadow-md rounded-md"
+  //         style={{ marginLeft: 10 }}
+  //         onClick={() => doXenditPayment()}
+  //       >
+  //         Pay with xendit
+  //       </button>
+  //     );
+  //   }
+  // };
 
   return (
     <>
@@ -170,7 +187,8 @@ const MyBookingPage = () => {
               My booking page
             </h3>
             <p className="text-base leading-none mt-4 text-gray-800">
-              Paid using credit card ending with <span className="font-semibold">8822</span>
+              Paid using credit card ending with{" "}
+              <span className="font-semibold">8822</span>
             </p>
             <div className="flex justify-center items-center w-full mt-8  flex-col space-y-4 ">
               <div className="flex md:flex-row justify-start items-start md:items-center  border border-gray-200 w-full">
@@ -199,7 +217,7 @@ const MyBookingPage = () => {
                     <p className="text-xl lg:text-2xl font-semibold leading-5 lg:leading-6 text-gray-800">
                       $28.00
                     </p>
-                    {btnPay()}
+                    {/* {btnPay()} */}
                   </div>
                 </div>
               </div>
@@ -210,30 +228,50 @@ const MyBookingPage = () => {
                   <p className="text-base font-semibold leading-4  text-gray-800">
                     Shipping Method
                   </p>
-                  <p className="text-sm leading-5 text-gray-600">Payment with xendit</p>
+                  <p className="text-sm leading-5 text-gray-600">
+                    Payment with xendit
+                  </p>
                 </div>
               </div>
               <div className="flex flex-col space-y-4 w-full">
                 <div className="flex justify-center items-center w-full space-y-4 flex-col border-gray-200 border-b pb-4">
                   <div className="flex justify-between  w-full">
-                    <p className="text-base leading-4 text-gray-800">Subtotal</p>
+                    <p className="text-base leading-4 text-gray-800">
+                      Subtotal
+                    </p>
                     <p className="text-base leading-4 text-gray-600">$56.00</p>
                   </div>
                   <div className="flex justify-between  w-full">
-                    <p className="text-base leading-4 text-gray-800">Discount </p>
+                    <p className="text-base leading-4 text-gray-800">
+                      Discount{" "}
+                    </p>
                     <p className="text-base leading-4 text-gray-600">-</p>
                   </div>
                   <div className="flex justify-between  w-full">
-                    <p className="text-base leading-4 text-gray-800">Shipping</p>
+                    <p className="text-base leading-4 text-gray-800">
+                      Shipping
+                    </p>
                     <p className="text-base leading-4 text-gray-600">-</p>
                   </div>
                 </div>
                 <div className="flex justify-between items-center w-full">
-                  <p className="text-base font-semibold leading-4 text-gray-800">Total</p>
-                  <p className="text-base font-semibold leading-4 text-gray-600">$36.00</p>
+                  <p className="text-base font-semibold leading-4 text-gray-800">
+                    Total
+                  </p>
+                  <p className="text-base font-semibold leading-4 text-gray-600">
+                    $36.00
+                  </p>
                 </div>
                 <div>
-                  <div style={{ textAlign: "center", padding: "10px" }}>
+                  <button
+                    className="py-5 focus:outline-none  focus:ring-offset-2  w-full text-base font-medium leading-4 bg-[#266c6b] hover:bg-[#0d423f] focus:ring-4 focus:ring-[#266c6b] text-white rounded-md"
+                    onClick={() => setModalOn(true)}
+                  >
+                    Verification
+                  </button>
+                  {modalOn && <Modal setModalOn={setModalOn} />}
+
+                  {/* <div style={{ textAlign: "center", padding: "10px" }}>
                     {captureVideo && modelsLoaded ? (
                       <div className="flex w-full justify-center items-center pt-1 md:pt-4  xl:pt-8 space-y-6 md:space-y-8 flex-col">
                         <button
@@ -279,7 +317,7 @@ const MyBookingPage = () => {
                     )
                   ) : (
                     <></>
-                  )}
+                  )} */}
                 </div>
               </div>
             </div>
